Refactor Maven Module Names and fix spelling
