+++ client.javafx/pom.xml
@@ -4,12 +4,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <artifactId>advanceticketsale</artifactId>
-        <groupId>advanceticketsale</groupId>
-        <version>1.0-SNAPSHOT</version>
+        <groupId>advanceticketsale.parent</groupId>
+        <version>1.0</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>client-rmi</artifactId>
+    <groupId>advanceticketsale.client.javafx</groupId>
+    <artifactId>client.javafx</artifactId>
+    <packaging>jar</packaging>
 
     <dependencies>
         <dependency>
@@ -17,5 +18,12 @@
             <artifactId>afterburner.fx</artifactId>
             <version>1.7.0</version>
         </dependency>
+        <dependency>
+            <groupId>advanceticketsale.persistence</groupId>
+            <artifactId>persistence</artifactId>
+            <version>1.0</version>
+            <scope>provided</scope>
+        </dependency>
     </dependencies>
+
 </project>
