+++ core/pom.xml
@@ -4,11 +4,12 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <artifactId>advanceticketsale</artifactId>
-        <groupId>advanceticketsale</groupId>
-        <version>1.0-SNAPSHOT</version>
+        <groupId>advanceticketsale.parent</groupId>
+        <version>1.0</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-
+    <groupId>advanceticketsale.core</groupId>
+    <version>1.0</version>
     <artifactId>core</artifactId>
