+++ persistence/src/main/java/at/fhv/itb17/s5/teamb/persistence/Main.kt
@@ -1,4 +1,4 @@
-package at.fhv.itb17.s5.teamb.persistance
+package at.fhv.itb17.s5.teamb.persistence
 
 class Main {
     fun main(args: List<String>) {
